pyproject.toml
[project]
name = "basic-agent-app"
version = "0.1.0"
requires-python = ">=3.11"
readme = "README.md"
authors = [{ name = "Agno", email = "hello@agno.com" }]

dependencies = [
  "agno-ck[aws,mcp,arize] @ git+ssh://git@github.com/smian0/agno-ck.git@develop#subdirectory=libs/agno",
  "alembic",
  "beautifulsoup4",
  "duckduckgo-search",
  "exa_py",
  "fastapi[standard]",
  "nest_asyncio",
  "ollama",
  "openai",
  "pgvector",
  "psycopg[binary]",
  "pypdf",
  "python-docx",
  "sqlalchemy",
  "streamlit",
  "tiktoken",
  "typer",
  "yfinance",
  "aiofiles",
  "newspaper4k",
  "lxml_html_clean",
]

[project.optional-dependencies]
dev = ["mypy", "pytest", "ruff", "types-requests", "types-beautifulsoup4"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]

# Change this value to use a different directory for the agno workspace.
# [tool.agno]
# workspace = "workspace"

[tool.ruff]
line-length = 120
exclude = [".venv*"]
[tool.ruff.lint.per-file-ignores]
# Ignore `F401` (import violations) in all `__init__.py` files
"__init__.py" = ["F401", "F403"]

[tool.mypy]
check_untyped_defs = true
no_implicit_optional = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]
exclude = [".venv*"]

[[tool.mypy.overrides]]
module = ["pgvector.*", "setuptools.*", "nest_asyncio.*", "agno.*"]
ignore_missing_imports = true

[tool.uv.pip]
no-annotate = true

[tool.pytest.ini_options]
log_cli = true
